+++ src/lib/log.js
@@ -1,13 +1,14 @@
 const initNow = Date.now();
 let lastNow = initNow;
 const log = (...text) => {
-  console.log( // eslint-disable-line no-console
+  // eslint-disable-next-line no-console
+  console.log(
     'profile',
     Date.now() - lastNow,
     Date.now() - initNow,
     ...text,
     initNow,
-  ); // eslint-disable-line no-console
+  );
   lastNow = Date.now();
 };
