+++ src/renderer/nvim/features/closeWindow.js
@@ -1,6 +1,9 @@
 import { nvim } from '../api';
 
-const { remote: { getCurrentWindow }, ipcRenderer } = global.require('electron');
+const {
+  remote: { getCurrentWindow },
+  ipcRenderer,
+} = global.require('electron');
 
 const currentWindow = getCurrentWindow();
 
@@ -9,7 +12,7 @@ const handleCloseWindow = () => {
 };
 
 const initCloseWindow = () => {
-  nvim().on('notification', (method) => {
+  nvim().on('notification', method => {
     if (method === 'vv:close_window') {
       currentWindow.webContents.send('quit');
     }
